chore: use express listen
